third_party/WebKit/LayoutTests/accessibility/textarea-line-for-index-expected.txt

This tests that the line number of a character offset is computed correctly for textarea elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axArea1.lineForIndex(character) is 0
FAIL axArea1.lineForIndex(character) should be 1. Was 0.
FAIL axArea1.lineForIndex(character) should be 2. Was 0.
FAIL axArea1.lineForIndex(character) should be 3. Was 0.
FAIL axArea1.lineForIndex(character) should be 4. Was 0.
FAIL axArea1.lineForIndex(character) should be 5. Was 0.
FAIL axArea1.lineForIndex(character) should be 6. Was 0.
PASS axArea1.lineForIndex(character) is 0
FAIL axArea1.lineForIndex(character) should be 1. Was 0.
FAIL axArea1.lineForIndex(character) should be 2. Was 0.
FAIL axArea1.lineForIndex(character) should be 3. Was 0.
FAIL axArea1.lineForIndex(character) should be 4. Was 0.
FAIL axArea1.lineForIndex(character) should be 5. Was 0.
FAIL axArea1.lineForIndex(character) should be 6. Was 0.
PASS axArea1.lineForIndex(character) is 0
FAIL axArea1.lineForIndex(character) should be 1. Was 0.
FAIL axArea1.lineForIndex(character) should be 2. Was 0.
FAIL axArea1.lineForIndex(character) should be 3. Was 0.
FAIL axArea1.lineForIndex(character) should be 4. Was 0.
FAIL axArea1.lineForIndex(character) should be 5. Was 0.
FAIL axArea1.lineForIndex(character) should be 6. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
